requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.7
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.3
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
ffmpeg-python==0.2.0
transformers==4.35.2
sentencepiece==0.1.99
regex==2023.10.3
tqdm==4.66.1
joblib==1.3.2
threadpoolctl==3.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
urllib3==2.0.7
# New packages for enhanced functionality
python-docx==1.1.0
twilio==8.10.3
flask-socketio==5.3.6
python-dotenv==1.0.0
# AWS Transcribe Streaming dependencies
boto3==1.35.36
botocore==1.35.36
websockets==12.0 

# AWS Integration
boto3==1.35.36
botocore==1.35.36
amazon-transcribe==0.6.2
python-dotenv==1.0.0 